requirements.txt
Flask==3.0.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
redis==5.0.1
